datasheetImport function
This allows the user to provide a link to a datasheet for a given function. This would likely be a schematic file of the function represented in kicad (for user friendliness) or a pdf printout. Similar concept can be made if a link to source code should be attached.

Key bit is that the link has to be assigned to a variable called 'Datasheet' as the funciton looks for that variable